Added new table definitions to GameData
I added the additional tables necessary to populate the full complement
of game data that we want to track. This will allow for the keeping of each
team's right, wrong, and skip status for each turn of each game. By keeping
data to this granularity, we get basic game data for free.