Filling in the game data functions.
Filled in the skeleton of Game.java such that game actions (team creation,
game creation, and scoring can be stored in a sqlite database). The next step
will be the creation of a game manager to run a game and find out what we
actually need out of this class.