Refactored the Game class.
I added some comments to the functions and changed some raw SQL construction
to the better, abstracted methods provided by the SDK. I'll take a look at
the xml-based creation at a later time. There is some raw SQL there that
should be changed to an 'insert' call, as well.