Heavily refactored the game manager.
In addition to adding a ton of comments that actually document our methods
within GameManager, I removed the cardPair type and started to use the newly
refactored and fully-featured Card class. This replacement allows the storage
of a reference to the card object instead of simply an id and a rws value.
Because of these changes, currentCards now becomes a linked list of Cards. This
method of storage is much more coherent than keeping a small private class
around. Now, everywhere you need a card, you use a Card object.