Implementing pausing properly.
Now, when the user pauses the application by either pulling up the menu or
leaving the application there is an overlay of "paused!" and the word and
badWords are hidden to the user. Additionally the right/wrong/skip button set
is disabled until the game is resumed by clicking on the pause button.