GameEnd now shows scores for all teams "properly"
Final scores for all participating teams are shown.  Scores for teams who are not entered will be hidden.  Table is now created dynamically without use of an adapter, so that columns for teams could be hidden on demand.  Put in a hack to set the width of the text views on each score row, as I couldn't figure out how to inflate the xml attributes I supplied to each TextView in GameEndRow.xml.  Will fix before pushing.