Fixing the Awarder and putting it into place at GameEnd.
This commit is entirely too large. In this I added the finishing touches to the
awarder class and added its functionality to GameEnd. This required a few index
checks in the Award queries in Game as well as a LIMIT statement on one of the
queries. Seems to work well, though.