Created SoundManager class to simpifly SFX calls.
SoundManager manages the SoundPool object and loads
assets in on creation, so we can load when the title screen
comes up and keep them loaded from game to game.  I also
added a turn ready sound.  We'll see how it works.

Cleaned up calls to playsound to route through SoundManager.