Centered up buttons on Title screen.
Huge ordeal.  Grouped labels with their buttons.  Added a fake
banner on left to hide labels animating in.  This allowed me to
anchor the labels with the buttons, for perfect centering.  In
order to do this, I had to remove glow on the buttons and mirror
the settings button, as the fake banner had to be edge on with
the buttons.

Centered the buttons by adding them to a linear layout with the
title and giving equal weight.  Menu items now evenly distribute
throughout the vertical area of the phone, which causes squashing
on the small phones.