Fixing bugs I introduced when creating the deck.
These were related to a bad initial offset and the lack of closure of cursors.
I also removed the threading of the load. It scares me. I think the best thing
to do would be to thread the load at splash the first time the app is run if we
really want to worry about the very first load of the words.