Beginning the threaded database load on the first invocation of the app.
This is done by creating a deck and threading loadWords. This will cause
the oncreate function to be called and start deck creation. We'll see if
this results in bugs down the line, but it should be fairly safe (unless
the user is very fast and/or the phone is very slow).