Fixing the back behavior on the turn activity.
This involved changing how we process cards and keep our cache of
already played cards. Instead of processing before adding to the
local cache, we process the cards as they are displayed to the user
this makes a lot more sense. The state of the code before this commit
was unclear and not well-reasoned. I think the movement of
functionality into the Deck class caused this lack of clarity. I'm
glad we got some time to revisit and refactor this. It's a LOT more
clear now.