Added new scoreboard to TurnSummary.
Had to rework handling of score to always have an up-to-date
score, rather than predicting it by adding in TurnScore when
setting the text on the scoreboard widgets.  So Turn needed
to set the score on time up, and when the user ends game.
TurnSummary had to ammend the score with each card ammendment.