Added "NotSet" Status icon and cleaned up art.
TurnSumRow now uses the generic white row end element. It
is colored programatically. Removed need for 4 pngs for each
status, but we did export specific status icons that are layered
on top of the row end element. Removed references to those in
GameManager (which probably shouldn't have had it anyways).
Added new button art for removing a card's status in CardReview.