Code cleanup and bugfix to shuffling card pull query.
Fixed issue where cards would get seen multiple times due to getting
timestamped the same as all others seen during the same turn. We fixed
this by doing a secondary sort on the query for cards that sorts on
number of times seen.

Removed some functions that weren't being used or could be consolidated,
and renamed variables that were confusing.