Improved memory management of MediaPlayer.
All activities now release the MediaPlayer when paused, except for Turn.
This will have the side-effect that the music will restart every time,
but it is better practice for managing system resources. I also now
create a MediaPlayer if Get() would have returned a NULL media player.
I believe this will fix many of the crash reports we've had.