Fixed bug where TurnSummary Next button incorrectly showed GameResults.
When playing scoreLimit and reaching the end of game condition on the final
team, modifying the cards so that the end of game condition was no longer
true did not refresh the Next button to read "Next Team" instead of "GameOver".
It also now reshows the score and point limit when this happens.