Altered App activity flow to Title -> GameSetup -> PP instead of T -> PP -> GS.
This required some changes to GameManager to allow it to store references to
the teams and game mode before starting a game. It also required the gameManager
be created in GameSetup for the first time, instead of in PackPurchase.

Reworked strings, buttons, and function names to be agnostic to the activity
flow.